examples/FalMeasurement/pyproject.toml
[tool.poetry]
name = "FalMeasurement"
version = "0.5.0"
package-mode = false
description = ""
authors = ["National Instruments"]

[tool.poetry.dependencies]
python = "^3.8"
ni-measurementlink-service = {version = "^1.3.0"}
PyVISA = "^1.13.0"
PyVISA-sim = "^0.5.1"
nidcpower = { version = ">=1.4.4", extras = ["grpc"] }
click = ">=7.1.2"
grpcio = "*"
python-decouple = ">=3.8"

[tool.poetry.group.dev.dependencies]
ni-python-styleguide = ">=0.4.1"
mypy = ">=1.0"
grpc-stubs = "^1.53"
# Uncomment to use prerelease dependencies.
# nidcpower = { git = "https://github.com/ni/nimi-python.git", subdirectory = "generated/nidcpower", extras=["grpc"] }
# ni-measurementlink-service = {path = "../..", develop = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "decouple.*",
    "hightime.*",
    "nidcpower.*",
]
ignore_missing_imports = true